% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{.DollarNames,ACTIONetExperiment-method}
\alias{.DollarNames,ACTIONetExperiment-method}
\title{Auto-completion for \code{\linkS4class{ACTIONetExperiment}} objects}
\usage{
\S4method{.DollarNames}{ACTIONetExperiment}(x, pattern = "")
}
\description{
Auto-completion for \code{\linkS4class{ACTIONetExperiment}} objects
}
