% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate.cells.from.archetypes.using.markers}
\alias{annotate.cells.from.archetypes.using.markers}
\title{Annotates cells by interpolating marker-based archetype annotations}
\usage{
annotate.cells.from.archetypes.using.markers(ace, marker.genes,
  rand.sample.no = 1000)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{marker.genes}{A list of lists (each a set of markers for a given cell type)}

\item{rand.sample.no}{Number of random permutations (default=1000)}
}
\value{
A named list: \itemize{
\item Labels: Inferred archetype labels
\item Labels.confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Annotates cells by interpolating marker-based archetype annotations
}
\examples{

data("curatedMarkers_human") # pre-packaged in ACTIONet
marker.genes = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
cell.annotations = annotate.cells.from.archetypes.using.markers(ace, marker.genes)
labels = cell.annotations$Labels
}
