% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate.cells.using.markers}
\alias{annotate.cells.using.markers}
\title{Directly inferring cell annotations from imputed gene expressions using permutation test}
\usage{
annotate.cells.using.markers(ace, marker.genes, rand.sample.no = 100,
  alpha_val = 0.9, thread_no = 8, imputation = "PageRank",
  data.slot = "logcounts")
}
\arguments{
\item{ace}{ACTIONet output object}

\item{marker.genes}{A list of lists (each a set of markers for a given cell type)}

\item{rand.sample.no}{Number of random permutations (default=1000)}

\item{alpha_val}{Random-walk parameter for gene imputation (if imputation = 'PageRank')}

\item{thread_no}{Number of parallel threads used for gene imputation (if imputation = 'PageRank')}

\item{imputation}{Gene imputation method. Can be either 'PageRank' (default) or 'archImpute'}
}
\value{
A named list: \itemize{
\item Labels: Inferred archetype labels
\item Labels.confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Directly inferring cell annotations from imputed gene expressions using permutation test
}
\examples{
data('curatedMarkers_human') # pre-packaged in ACTIONet
marker.genes = curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
arch.annot = annotate.cells.using.markers(ace, marker.genes = marker.genes)
cell.labels = arch.annot$Labels
}
