% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{annotate.profile.using.markers}
\alias{annotate.profile.using.markers}
\title{Annotate arbitary feature score matrix using known marker genes
(It uses permutation test on cluster specificity scores)}
\usage{
annotate.profile.using.markers(profile, markers)
}
\arguments{
\item{profile}{An arbitrary matrix with rows corresponding to features and columns to any given annotation/grouping of cells}

\item{markers}{A list of lists (each a set of markers for a given cell type)}
}
\value{
A named list: \itemize{
\item Labels: Inferred archetype labels
\item Labels.confidence: Confidence of inferred labels
\item Enrichment: Full enrichment matrix
}
}
\description{
Annotate arbitary feature score matrix using known marker genes
(It uses permutation test on cluster specificity scores)
}
