% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{archetypeFeatureSpecificity}
\alias{archetypeFeatureSpecificity}
\title{Computes feature specificity scores for each archetype.}
\usage{
archetypeFeatureSpecificity(
  ace,
  assay_name = "logcounts",
  footprint_slot = "archetype_footprint",
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Computes feature specificity scores for each archetype.
}
