% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{assess.geneset.enrichment.from.scores}
\alias{assess.geneset.enrichment.from.scores}
\title{PErforms geneset enrichment analysis on arbitrary gene scores}
\usage{
assess.geneset.enrichment.from.scores(scores, associations, L = 1000)
}
\arguments{
\item{scores}{Gene scores}

\item{associations}{Either a genes x pathways membership matrix, or a set of genesets}

\item{L}{Maximum length of the top-ranked genes to consider}
}
\value{
Matrix pathway x cell type/states
}
\description{
PErforms geneset enrichment analysis on arbitrary gene scores
}
\examples{
data("gProfilerDB_human")
associations = gProfilerDB_human$SYMBOL$WP
scores = rowFactors(ace)$archetype_gene_specificity
Geneset.enrichments = assess.geneset.enrichment.from.scores(scores, associations)
}
