% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{assess.geneset.enrichment}
\alias{assess.geneset.enrichment}
\title{Performs geneset enrichment analysis on archetypes}
\usage{
assess.geneset.enrichment(ace, associations, L = 1000,
  specificity.slot = "unified_feature_specificity")
}
\arguments{
\item{ace}{ACTIONetExperiment (ACE) output object}

\item{associations}{Either a genes x pathways membership matrix, or a set of genesets}

\item{L}{Maximum length of the top-ranked genes to consider}
}
\value{
Matrix pathway x cell type/states
}
\description{
Performs geneset enrichment analysis on archetypes
}
\examples{
data("gProfilerDB_human")
associations = gProfilerDB_human$SYMBOL$WP
Geneset.enrichments = assess.geneset.enrichment.from.archetypes(ace, associations)
}
