% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_ACTIONet}
\alias{build_ACTIONet}
\title{Builds an interaction network from the multi-level archetypal decompositions}
\usage{
build_ACTIONet(H_stacked, density = 1, thread_no = 0L,
  mutual_edges_only = TRUE)
}
\arguments{
\item{H_stacked}{Output of the prune_archetypes() function.}

\item{density}{Overall density of constructed graph. The higher the density, the more edges are retained (default = 1.0).}

\item{thread_no}{Number of parallel threads (default = 0).}

\item{mutual_edges_only}{Symmetrization strategy for nearest-neighbor edges. 
If it is true, only mutual-nearest-neighbors are returned (default=TRUE).}
}
\value{
G Adjacency matrix of the ACTIONet graph.
}
\description{
Builds an interaction network from the multi-level archetypal decompositions
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
\\tG = build_ACTIONet(prune.out$H_stacked)
}
