% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellMarkerDB_mouse}
\alias{CellMarkerDB_mouse}
\title{Pre-processed cell type markers from the CellMarker database (Mouse)}
\format{:
  A gene x cell-type indicator matrix in the sparseMatrix format (rownames are gene symbols)}
\source{
\url{http://biocc.hrbmu.edu.cn/CellMarker/}
}
\usage{
CellMarkerDB_mouse
}
\description{
Pre-processed cell type markers from the CellMarker database (Mouse)
}
\keyword{datasets}
