% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{cluster.graph}
\alias{cluster.graph}
\title{A wrapper function for the leiden algorithm}
\usage{
cluster.graph(G, resolution_parameter = 0.5, initial_clusters = NULL, seed = 0)
}
\arguments{
\item{G}{Adjacency matrix of the input graph}

\item{resolution_parameter}{Resolution of the clustering.
The higher the resolution, the more clusters we will get (default=0.5).}

\item{initial_clusters}{Used as the inital clustering}

\item{seed}{Random seed}
}
\value{
ace with added annotation
}
\description{
A wrapper function for the leiden algorithm
}
\examples{
clusters <- cluster.graph(G, 1.0)
}
