% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{clusterFeatureSpecificity}
\alias{clusterFeatureSpecificity}
\title{Computes feature specificity scores for each cluster.}
\usage{
clusterFeatureSpecificity(
  ace = NULL,
  S = NULL,
  cluster_attr = NULL,
  output_prefix = NULL,
  assay_name = "logcounts",
  return_raw = FALSE
)
}
\description{
Computes feature specificity scores for each cluster.
}
