% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_ace.R
\docType{methods}
\name{colEmbeddings}
\alias{colEmbeddings}
\alias{colEmbeddings<-,ACTIONetExperiment-method}
\alias{rowEmbeddings<-,ACTIONetExperiment-method}
\title{Get column embeddings}
\usage{
colEmbeddings(x, ...)

\S4method{colEmbeddings}{ACTIONetExperiment}(x) <- value

\S4method{rowEmbeddings}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices

List of matrices
}
\description{
Get column embeddings

Set column-associated embeddings

Set row-associated embeddings
}
