% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colFactors<-}
\alias{colFactors<-}
\title{Set colFactors}
\usage{
colFactors(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set colFactors
}
