% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_ace.R, R/set_ace.R
\docType{methods}
\name{colFactors}
\alias{colFactors}
\alias{colFactors,ACTIONetExperiment-method}
\alias{colFactors<-,ACTIONetExperiment-method}
\title{Get colFactors}
\usage{
colFactors(x, ...)

\S4method{colFactors}{ACTIONetExperiment}(x, withDimnames = TRUE)

\S4method{colFactors}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices

List of matrices
}
\description{
Get colFactors

Get column-associated factors

Set column-associated factors
}
