% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colMapMeta<-}
\alias{colMapMeta<-}
\title{Set colMapMeta}
\usage{
colMapMeta(obj, ...) <- value
}
\arguments{
\item{obj}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set colMapMeta
}
