% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_ace.R
\docType{methods}
\name{colMapMeta}
\alias{colMapMeta}
\alias{colMapMeta<-,ACTIONetExperiment-method}
\title{Get colMapTypes}
\usage{
colMapMeta(x, ...)

\S4method{colMapMeta}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices
}
\description{
Get colMapTypes

Set column-associated factor annotations
}
