% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_ace.R, R/set_ace.R
\docType{methods}
\name{colMaps}
\alias{colMaps}
\alias{colMaps,ACTIONetExperiment-method}
\alias{colMaps<-,ACTIONetExperiment-method}
\title{Get colMaps}
\usage{
colMaps(x, ...)

\S4method{colMaps}{ACTIONetExperiment}(x, all = T)

\S4method{colMaps}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices

List of matrices
}
\description{
Get colMaps

Get column-associated factors

Set column-associated factors
}
