% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colMapTypes<-}
\alias{colMapTypes<-}
\title{Set colMapTypes}
\usage{
colMapTypes(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set colMapTypes
}
