% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colMapTypes}
\alias{colMapTypes}
\title{Get colMapTypes}
\usage{
colMapTypes(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get colMapTypes
}
