% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_ace.R, R/set_ace.R
\docType{methods}
\name{colNets}
\alias{colNets}
\alias{colNets,ACTIONetExperiment-method}
\alias{colNets<-,ACTIONetExperiment-method}
\title{Get colNets}
\usage{
colNets(x, ...)

\S4method{colNets}{ACTIONetExperiment}(x)

\S4method{colNets}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of adjacency matrices

List of adjacency matrices
}
\description{
Get colNets

Get column-associated networks

Set column-associated networks
}
