% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colReductions<-}
\alias{colReductions<-}
\title{Set column reductions}
\usage{
colReductions(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set column reductions
}
