% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colReductions}
\alias{colReductions}
\title{Get column reductions}
\usage{
colReductions(x, ...)
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get column reductions
}
