% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{compute.cluster.feature.specificity}
\alias{compute.cluster.feature.specificity}
\title{Computes feature (i.e. gene) specificity scores for each cluster}
\usage{
compute.cluster.feature.specificity(
  ace,
  cluster_attr,
  output_prefix,
  assay_name = "logcounts",
  features_use = NULL
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{cluster_attr}{Vector of length NCOL(ace) or column name of colData(ace) containing cluster labels.}

\item{output_prefix}{String to prefix names of rowMaps(ace) where output is stored.}

\item{assay_name}{Name of assay to be used. (default="logcounts")}

\item{features_use}{A vector of features of length NROW(ace) or the name of a column of rowData(ace) containing the feature names. Default uses rownames(ace). (default=NULL)}
}
\value{
`ACE` object with specificity scores of each cluster added to rowMaps(ace) as a matrix with name defined by output_prefix
}
\description{
Computes feature (i.e. gene) specificity scores for each cluster
}
\examples{
ace <- compute.cluster.feature.specificity(ace, ace$clusters, "cluster_specificity_scores")
}
