% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_AA_coreset}
\alias{compute_AA_coreset}
\title{Computes a coreset for archetypal analysis
Ref: Coresets for Archetypal Analysis (http://papers.neurips.cc/paper/8945-coresets-for-archetypal-analysis)}
\usage{
compute_AA_coreset(S, m = 0L)
}
\arguments{
\item{S}{Input matrix (e.g., gene x cell)}

\item{m}{Number of samples (or 0, to be automatically identified)}

\item{seed}{Random seed}
}
\value{
clusters Assignment vector of samples to clusters
}
\description{
Computes a coreset for archetypal analysis
Ref: Coresets for Archetypal Analysis (http://papers.neurips.cc/paper/8945-coresets-for-archetypal-analysis)
}
\examples{
coreset = compute_AA_coreset(S, 1000)
}
