% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_archetype_core_centrality}
\alias{compute_archetype_core_centrality}
\title{Compute coreness of subgraph vertices induced by each archetype}
\usage{
compute_archetype_core_centrality(G, sample_assignments)
}
\arguments{
\item{G}{Input graph}

\item{sample_assignments}{Archetype discretization (output of unify_archetypes())}
}
\value{
cn core-number of each graph node
}
\description{
Compute coreness of subgraph vertices induced by each archetype
}
\examples{
G = colNets(ace)$ACTIONet
assignments = ace$archetype.assignment
connectivity = compute_core_number(G, assignments)
}
