% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_archetype_feature_specificity}
\alias{compute_archetype_feature_specificity}
\title{Compute feature specificity (from archetype footprints)}
\usage{
compute_archetype_feature_specificity(S, H)
}
\arguments{
\item{S}{Input matrix (sparseMatrix)}

\item{H}{A soft membership matrix - Typically H_unified from the unify_archetypes() function.}
}
\value{
A list with the over/under-logPvals
}
\description{
Compute feature specificity (from archetype footprints)
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
G = build_ACTIONet(prune.out$H_stacked)
unification.out = unify_archetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
S.norm = renormalize_input_matrix(S, cell.clusters)
logPvals.list = compute_archetype_feature_specificity(S.norm, unification.out$H_unified)
specificity.scores = logPvals.list$upper_significance
}
