% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_archetype_feature_specificity_bin}
\alias{compute_archetype_feature_specificity_bin}
\title{Compute feature specificity (from archetype footprints and binary input)}
\usage{
compute_archetype_feature_specificity_bin(S, H)
}
\arguments{
\item{S}{Input matrix (sparseMatrix - binary)}

\item{H}{A soft membership matrix - Typically H_unified from the unify_archetypes() function.}
}
\value{
A list with the over/under-logPvals
}
\description{
Compute feature specificity (from archetype footprints and binary input)
}
\examples{
\\tlogPvals.list = compute_archetype_feature_specificity_bin(S.bin, unification.out$H_unified)
specificity.scores = logPvals.list$upper_significance
}
