% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_feature_specificity}
\alias{compute_feature_specificity}
\title{Compute feature specificity (discriminative scores)}
\usage{
compute_feature_specificity(S, H)
}
\arguments{
\item{S}{Input matrix}

\item{H}{A soft membership matrix - Typically H_unified from the unify_archetypes() function.}
}
\value{
A list with the over/under-logPvals
}
\description{
Compute feature specificity (discriminative scores)
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
G = build_ACTIONet(prune.out$H_stacked)
unification.out = unify_archetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
logPvals.list = renormalize_input_matrix(S, unification.out$H_unified)
}
