% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_grouped_rowsums}
\alias{compute_grouped_rowsums}
\title{Aggregate matrix wiithin groups}
\usage{
compute_grouped_rowsums(S, sample_assignments)
}
\arguments{
\item{S}{matrix of type "dMatrix"}

\item{sample_assignments}{Vector of column groupings. Group labels must be continuous integers or coercible to such.}
}
\value{
S matrix with columns of values aggregated within each group of sample_assignments
}
\description{
Aggregate matrix wiithin groups
}
