% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_network_diffusion}
\alias{compute_network_diffusion}
\title{Computes network diffusion over a given network, starting with an arbitrarty set of initial scores}
\usage{
compute_network_diffusion(G, X0, thread_no = 0L, alpha = 0.85,
  max_it = 3L)
}
\arguments{
\item{G}{Input graph}

\item{X0}{Matrix of initial values per diffusion (ncol(G) == nrow(G) == ncol(X0))}

\item{thread_no}{Number of parallel threads (default=0)}

\item{alpha}{Random-walk depth ( between [0, 1] )}

\item{max_it}{PageRank iterations}
}
\value{
Matrix of diffusion scores
}
\description{
Computes network diffusion over a given network, starting with an arbitrarty set of initial scores
}
\examples{
G = colNets(ace)$ACTIONet
gene.expression = Matrix::t(logcounts(ace))[c('CD19', 'CD14', 'CD16'), ]
smoothed.expression = compute_network_diffusion(G, gene.expression)
}
