% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_network_diffusion_direct}
\alias{compute_network_diffusion_direct}
\title{Computes network diffusion over a given network, starting with an arbitrarty set of initial scores (direct approach)}
\usage{
compute_network_diffusion_direct(G, X0, thread_no = 0L, alpha = 0.85)
}
\arguments{
\item{G}{Input graph}

\item{X0}{Matrix of initial values per diffusion (ncol(G) == nrow(G) == ncol(X0))}

\item{thread_no}{Number of parallel threads (default=0)}

\item{alpha}{Random-walk depth ( between [0, 1] )}

\item{max_it}{PageRank iterations}
}
\value{
Matrix of diffusion scores
}
\description{
Computes network diffusion over a given network, starting with an arbitrarty set of initial scores (direct approach)
}
\examples{
G = colNets(ace)$ACTIONet
gene.expression = Matrix::t(logcounts(ace))[c("CD19", "CD14", "CD16"), ]
smoothed.expression = compute_network_diffusion_direct(G, gene.expression)
}
