% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_pseudo_bulk_per_ind}
\alias{compute_pseudo_bulk_per_ind}
\title{Computes pseudobulk profiles (groups[k1] x individuals[k2])}
\usage{
compute_pseudo_bulk_per_ind(S, sample_assignments, individuals)
}
\arguments{
\item{S}{Input matrix ('sparseMatrix')}

\item{sample_assignments}{Any primary grouping - typically based on cell type/state (it has to be in {1, ..., k1})}

\item{individuals}{Any Secondary grouping - typically corresponds to individuals (it has to be in {1, ..., k2})}
}
\value{
A list of pseudobulk profile, where each entry is matrix corresponding to one cell type/state
}
\description{
Computes pseudobulk profiles (groups[k1] x individuals[k2])
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
\\tG = build_ACTIONet(prune.out$H_stacked)
unification.out = unify_archetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
\\tpbs.list = compute_pseudo_bulk(S, cell.clusters, sce$individuals)
}
