% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_sparse_network_diffusion}
\alias{compute_sparse_network_diffusion}
\title{Computes sparse network diffusion over a given network, starting with an arbitrarty set of initial scores}
\usage{
compute_sparse_network_diffusion(
  G,
  X0,
  alpha = 0.85,
  rho = 1e-04,
  epsilon = 0.001,
  max_iter = 20L
)
}
\arguments{
\item{G}{Input graph}

\item{X0}{Matrix of initial values per diffusion (ncol(G) == nrow(G) == ncol(X0))}

\item{alpha}{Random-walk depth ( between [0, 1] )}

\item{rho}{Sparsity controling parameter}

\item{epsilon, max_it}{Conditions on the length of diffusion}
}
\value{
Matrix of sparse diffusion scores
}
\description{
Computes sparse network diffusion over a given network, starting with an arbitrarty set of initial scores
}
\examples{
G = colNets(ace)$ACTIONet
gene.expression = Matrix::t(logcounts(ace))[c("CD19", "CD14", "CD16"), ]
smoothed.expression = compute_sparse_network_diffusion(G, gene.expression)
}
