% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{convert.sce.rownames}
\alias{convert.sce.rownames}
\title{Translates rownames() of the `sce` object for mouse and human datasets.
A typical use-case is when input ids are in ENSEMBL format, but user is interested to work with gene symbols.}
\usage{
convert.sce.rownames(sce, from = "ENSEMBL", to = "SYMBOL", species = "human")
}
\arguments{
\item{sce}{Input `sce` object}

\item{from}{Source annotation (default="ENSEMBL")}

\item{to}{Target annotation (default="SYMBOL")}

\item{species}{Either "mouse" or "human" (default="human")}
}
\value{
`SingleCellExeriment` object with renamed rows
}
\description{
Translates rownames() of the `sce` object for mouse and human datasets.
A typical use-case is when input ids are in ENSEMBL format, but user is interested to work with gene symbols.
}
\examples{
sce = import.sce.from.CDS(monocle_cds)
}
