% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{correct.cell.annotations}
\alias{correct.cell.annotations}
\title{Uses a variant of the label propagation algorithm to correct likely noisy labels}
\usage{
correct.cell.annotations(
  ace,
  initial_labels,
  algorithm = "lpa",
  iters = 3,
  lambda = 0,
  sig_threshold = 3,
  net_slot = "ACTIONet"
)
}
\arguments{
\item{ace}{Input results to be clustered
(alternatively it can be the ACTIONet igraph object)}

\item{initial_labels}{Annotations to correct with missing values (NA) in it.
It can be either a named annotation (inside ace$annotations) or a label vector.}

\item{iters}{How many iterative rounds of correction/inference should be performed (default=3)}

\item{LFR.threshold}{How aggressively to update labels. The smaller the value, the more labels will be changed (default=2)}

\item{double.stochastic}{Whether to densify adjacency matrix before running label propagation (default=FALSE).}
}
\value{
ace with updated annotations added to ace$annotations
}
\description{
Uses a variant of the label propagation algorithm to correct likely noisy labels
}
\examples{
ace <- add.cell.annotations(ace, cell.labels, "input_annotations")
ace <- correct.cell.annotations(ace, "input_annotations", "updated_annotations")
}
