% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decomp.ACTION_MR}
\alias{decomp.ACTION_MR}
\title{Run ACTION_decomposition_MR}
\usage{
decomp.ACTION_MR(
  ace = NULL,
  S_r = NULL,
  k_min = 2,
  k_max = 30,
  min_specificity_z_thresh = -3,
  min_cells_per_arch = 2,
  unification_violation_threshold = 0,
  max_iter = 50,
  thread_no = 0,
  reduction_slot = "ACTION",
  return_raw = FALSE,
  seed = 0
)
}
\description{
Run ACTION_decomposition_MR
}
