% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal_main.R
\name{.run.ACTIONMR.ace}
\alias{.run.ACTIONMR.ace}
\title{Run ACTION_decomposition_MR}
\usage{
.run.ACTIONMR.ace(
  ace,
  k_min = 2,
  k_max = 30,
  specificity_th = -3,
  min_cells_per_arch = 2,
  unification_backbone_density = 0.5,
  unification_resolution = 1,
  unification_min_cluster_size = 3,
  max_iter = 50,
  min_delta = 1e-300,
  thread_no = 0,
  unified_suffix = "unified",
  footprint_slot_name = "assigned_archetype",
  reduction_slot = "ACTION"
)
}
\description{
Run ACTION_decomposition_MR
}
