% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal_main.R
\name{.run.pruneArchetypes}
\alias{.run.pruneArchetypes}
\title{Prune nonspecific and/or unreliable archetypes}
\usage{
.run.pruneArchetypes(
  C_trace,
  H_trace,
  ace = NULL,
  min_specificity_z_thresh = -3,
  min_cells_per_arch = 2,
  return_raw = FALSE
)
}
\description{
Prune nonspecific and/or unreliable archetypes
}
