% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal_main.R
\name{.run.unifyArchetypes}
\alias{.run.unifyArchetypes}
\title{Identiy equivalent classes of archetypes and group them together}
\usage{
.run.unifyArchetypes(
  ace = NULL,
  S_r = NULL,
  C_stacked = NULL,
  H_stacked = NULL,
  reduction_slot = "ACTION",
  C_stacked_slot = "C_stacked",
  H_stacked_slot = "H_stacked",
  unified_suffix = "unified",
  violation_threshold = 0,
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Identiy equivalent classes of archetypes and group them together
}
