% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EnrichrDB}
\alias{EnrichrDB}
\title{Enrichr functional pathways}
\format{:
  A list of genesets}
\source{
\url{https://amp.pharm.mssm.edu/Enrichr/}
}
\usage{
EnrichrDB
}
\description{
Enrichr functional pathways
}
\keyword{datasets}
