% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FengSVD}
\alias{FengSVD}
\title{Computes SVD decomposition}
\usage{
FengSVD(A, dim, iters = 5L, seed = 0L)
}
\arguments{
\item{A}{Input matrix ("sparseMatrix")}

\item{dim}{Dimension of SVD decomposition}

\item{iters}{Number of iterations (default=5)}

\item{seed}{Random seed (default=0)}
}
\value{
A named list with U, sigma, and V components
}
\description{
This is direct implementation of the randomized SVD algorithm for sparse matrices:
Xu Feng, Yuyang Xie, and Yaohang Li, "Fast Randomzied SVD for Sparse Data," in Proc. the 10th Asian Conference on Machine Learning (ACML), Beijing, China, Nov. 2018.
}
\examples{
A = randn(100, 20)
SVD.out = FengSVD(A, dim = 2)
U = SVD.out$U
}
