% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.annotation.markers}
\alias{get.annotation.markers}
\title{Report the top-ranked marker genes for each arbitrary annotation}
\usage{
get.annotation.markers(ACTIONet.out, annotation.name, top.genes = 20)
}
\arguments{
\item{ACTIONet.out}{Output list to be pruned}

\item{top.genes}{Number of genes to return}
}
\value{
A data.frame with each column representing the top-ranked genes for an archetype.
}
\description{
Report the top-ranked marker genes for each arbitrary annotation
}
\examples{
ACTIONet.out.pruned = get.archetype.markers(ACTIONet.out, 10)
}
