% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HalkoSVD}
\alias{HalkoSVD}
\title{Computes SVD decomposition}
\usage{
HalkoSVD(A, dim, iters = 5L, seed = 0L)
}
\arguments{
\item{A}{Input matrix ('sparseMatrix')}

\item{dim}{Dimension of SVD decomposition}

\item{iters}{Number of iterations (default=5)}

\item{seed}{Random seed (default=0)}
}
\value{
A named list with U, sigma, and V components
}
\description{
This is direct implementation of the randomized SVD algorithm:
XFrom: N Halko, P. G Martinsson, and J. A Tropp. Finding structure with randomness: Probabilistic algorithms for constructing approximate matrix decompositions. Siam Review, 53(2):217-288, 2011.
}
\examples{
A = randn(100, 20)
SVD.out = HalkoSVD(A, dim = 2)
U = SVD.out$U
}
