% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.ace.from.10X.generic}
\alias{import.ace.from.10X.generic}
\title{Imports data from a 10X experiment folder and constructs an `SingleCellExeriment` object}
\usage{
import.ace.from.10X.generic(
  input_path,
  sub_path = "filtered_gene_bc_matrices",
  mtx_file = "matrix.mtx.gz",
  feature_annotations = "features.tsv.gz",
  sample_annotations = "barcodes.tsv.gz",
  sep = "\\t",
  prefilter = FALSE,
  min.cell.frac.per.gene = 0.005,
  min.genes.per.cell = 500,
  use.names = TRUE
)
}
\arguments{
\item{input_path}{Folder containing input files}

\item{mtx_file}{Count file in Matrix Market format (default="matrix.mtx.gz")}

\item{feature_annotations}{Table of the same size as number of rows in the count matrix (default="features.tsv.gz")}

\item{sample_annotations}{Table of the same size as number of columns in the count matrix (default="barcodes.tsv.gz")}

\item{sep}{Column-separator used in the row/column annotations files (default='\\t')}

\item{prefilter}{Whether to prefilter genes/cells based on the counts.mat}

\item{min.cell.frac.per.gene}{Minimum fraction of cells capturing a gene for it to be retained, if prefilter=TRUE (default=0.005)}

\item{min.genes.per.cell}{Minimum number of required captured genes per cell, if prefilter=TRUE (default=500)}
}
\value{
`SingleCellExeriment` object
}
\description{
Imports data from a 10X experiment folder and constructs an `SingleCellExeriment` object
}
\examples{
ace = import.ace.from.10X.generic(input_path, prefilter=TRUE)
}
