% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.sce.from.AnnData}
\alias{import.sce.from.AnnData}
\title{Constructs an `SingleCellExeriment` object from AnnData
Please refer to: https://satijalab.org/seurat/v3.0/conversion_vignette.html}
\usage{
import.sce.from.AnnData(AnnData.fname)
}
\arguments{
\item{AnnData.fname}{Path to the AnnData file}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from AnnData
Please refer to: https://satijalab.org/seurat/v3.0/conversion_vignette.html
}
\examples{
sce = import.sce.from.AnnData(file_name)
}
