% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.sce.from.CDS}
\alias{import.sce.from.CDS}
\title{Constructs an `SingleCellExeriment` object from CDS format in Monocle}
\usage{
import.sce.from.CDS(monocle_cds)
}
\arguments{
\item{monocle_cds}{CDS object}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from CDS format in Monocle
}
\examples{
sce = import.sce.from.CDS(monocle_cds)
}
