% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.sce.from.count.matrix}
\alias{import.sce.from.count.matrix}
\title{Constructs an `SingleCellExeriment` object from count matrix, gene names, and sample_annotations}
\usage{
import.sce.from.count.matrix(counts, gene.names,
  sample_annotations = NULL, prefilter = FALSE,
  min.cell.frac.per.gene = 0.005, min.genes.per.cell = 500)
}
\arguments{
\item{counts}{Matrix of counts}

\item{gene.names}{A list of gene names to annotate columns of the count matrix}

\item{sample_annotations}{Optional table of annotations for samples (columns of the count matrix)}

\item{prefilter}{Whether to prefilter genes/cells based on the counts}

\item{min.cell.frac.per.gene}{Minimum fraction of cells capturing a gene for it to be retained, if prefilter=TRUE (default=0.005)}

\item{min.genes.per.cell}{Minimum number of required captured genes per cell, if prefilter=TRUE (default=500)}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from count matrix, gene names, and sample_annotations
}
\examples{
sce = import.sce.from.count.matrix(counts.mat, gene_names, prefilter=TRUE)
}
