% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.sce.from.loom}
\alias{import.sce.from.loom}
\title{Constructs an `SingleCellExeriment` object from AnnData
This function depends on sceasy (https://github.com/cellgeni/sceasy) and 
LoomExperiment(BioC) for file conversion.}
\usage{
import.sce.from.loom(fname)
}
\arguments{
\item{fname}{Path to the AnnData file}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from AnnData
This function depends on sceasy (https://github.com/cellgeni/sceasy) and 
LoomExperiment(BioC) for file conversion.
}
\examples{
sce = import.sce.from.loom(file_name)
}
