% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import.sce.from.table}
\alias{import.sce.from.table}
\title{Constructs an `SingleCellExeriment` object from a full count matrix file}
\usage{
import.sce.from.table(
  fname,
  sep = "\\t",
  prefilter = FALSE,
  min.cell.frac.per.gene = 0.001,
  min.genes.per.cell = 300
)
}
\arguments{
\item{fname}{Full path to the count matrix file}

\item{sep}{Column-separator used in count matrix file (default='\\t')}

\item{prefilter}{Whether to prefilter genes/cells based on the counts}

\item{min.cell.frac.per.gene}{Minimum fraction of cells capturing a gene for it to be retained, if prefilter=TRUE (default=0.005)}

\item{min.genes.per.cell}{Minimum number of required captured genes per cell, if prefilter=TRUE (default=500)}
}
\value{
`SingleCellExeriment` object
}
\description{
Constructs an `SingleCellExeriment` object from a full count matrix file
}
\examples{
sce = import.sce.from.table(file_name, prefilter=TRUE)
}
