% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{impute.genes.using.ACTIONet}
\alias{impute.genes.using.ACTIONet}
\title{Gene expression imputation using network diffusion.}
\usage{
impute.genes.using.ACTIONet(
  ACTIONet.out,
  sce,
  genes,
  alpha_val = 0.85,
  thread_no = 8,
  prune = FALSE,
  diffusion_iters = 5,
  sce.data.attr = "logcounts"
)
}
\arguments{
\item{ACTIONet.out}{Input results to be clustered
(alternatively it can be the ACTIONet igraph object)}

\item{sce}{Input single-cell profile witht the normalized counts}

\item{genes}{The list of genes to perform imputation for.}

\item{alpha_val}{Depth of diffusion between (0, 1).
The larger it is, the deeper the diffusion, which results in less nonzeros (default = 0.85).}

\item{thread_no}{Number of parallel threads}

\item{prune}{Whether to post-process diffusion results to remove false-positives (default = FALSE)}

\item{diffusion_iters}{Number of diffusion iterations (default = 5)}

\item{sce.data.attr}{Slot in the sce object with normalized counts.}
}
\value{
Imputed gene expression matrix. Column names are set with imputed genes names and rows are cells.
}
\description{
Gene expression imputation using network diffusion.
}
\examples{
imputed.genes = impute.genes.using.ACTIONet(ACTIONet.out, sce, c("MOBP", "C3", "GAD1", "VCAN"))
plot.ACTIONet.gradient(ACTIONet.out, imputed.genes[, 1])
}
