% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{layout_ACTIONet}
\alias{layout_ACTIONet}
\title{Performs stochastic force-directed layout on the input graph (ACTIONet)}
\usage{
layout_ACTIONet(G, S_r, compactness_level = 50L, n_epochs = 500L,
  thread_no = 0L)
}
\arguments{
\item{G}{Adjacency matrix of the ACTIONet graph}

\item{S_r}{Reduced kernel matrix (is used for reproducible initialization).}

\item{compactness_level}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50)}

\item{n_epochs}{Number of epochs for SGD algorithm (default=100).}

\item{thread_no}{Number of threads (default = 0).}
}
\value{
A named list \itemize{
\item coordinates 2D coordinates of vertices.
\item coordinates_3D 3D coordinates of vertices.
\item colors De novo color of nodes inferred from their 3D embedding.
}
}
\description{
Performs stochastic force-directed layout on the input graph (ACTIONet)
}
\examples{
G = build_ACTIONet(prune.out$H_stacked)
vis.out = layout_ACTIONet(G, S_r)
}
