% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{logcounts<-}
\alias{logcounts<-}
\title{Set logcounts assay}
\usage{
logcounts(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set logcounts assay
}
