% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Monaco2019}
\alias{Monaco2019}
\title{Reference profile for cell sorted PBMC}
\format{:
  SingleCellExperiment object with rowData, colData, and metadata inside}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/30726743}
}
\usage{
Monaco2019
}
\description{
Reference profile for cell sorted PBMC
}
\keyword{datasets}
