% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MWM_hungarian}
\alias{MWM_hungarian}
\title{Computes the maximum-weight bipartite graph matching}
\usage{
MWM_hungarian(G)
}
\arguments{
\item{G}{Adjacency matrix of the input graph}
}
\value{
G_matched An adjacency matrix with a maximum of one nonzero entry on rows/columns
}
\description{
Computes the maximum-weight bipartite graph matching
}
\examples{
G_matched = MWM_hungarian(G)
}
