% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{normcounts<-}
\alias{normcounts<-}
\title{Set normcounts assay}
\usage{
normcounts(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set normcounts assay
}
