% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{normcounts}
\alias{normcounts}
\title{Get normcounts assay}
\usage{
normcounts(x, ...)
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get normcounts assay
}
