% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{old.plot.ACTIONet}
\alias{old.plot.ACTIONet}
\title{Main ACTIONet plotting functions}
\usage{
old.plot.ACTIONet(
  ace,
  labels = NULL,
  transparency.attr = NULL,
  trans.z.threshold = -0.5,
  trans.fact = 1.5,
  node_size = 0.1,
  CPal = CPal20,
  add.text = TRUE,
  suppress.legend = TRUE,
  legend.pos = "bottomright",
  title = "",
  border.contrast.factor = 0.1,
  add.backbone = FALSE,
  arch.size.factor = 3,
  coordinate_slot = "ACTIONet2D"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to be projected on the ACTIONet plot}

\item{transparency.attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans.z.threshold, }{trans.fact Control the effect of transparency mapping}

\item{node_size}{Size of nodes in the ACTIONet plot}

\item{CPal}{Color palette (named vector or a name for a given known palette)}

\item{add.text}{Whether or not to add labels on the top of ACTIONet plot}

\item{suppress.legend}{Whether to suppress legend or include one}

\item{legend.pos}{If suppress.legend == F, where should we put the legend}

\item{title}{Main title of the plot}

\item{border.contrast.factor}{How much the node and its border should contrast}

\item{coordinate_slot}{Entry in colMaps(ace) containing the plot coordinates (default:'ACTIONet2D')}

\item{add.states}{Whether or not to include interpolated cell state positions in the plot}
}
\value{
Visualized ACTIONet
}
\description{
Main ACTIONet plotting functions
}
\examples{
ace = run.ACTIONet(sce)
plot.ACTIONet(ace, ace$assigned_archetype, transparency.attr = ace$node_centrality)
}
