% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PCA2ACTIONred_full}
\alias{PCA2ACTIONred_full}
\title{Computes reduced kernel matrix for a given (single-cell) profile and prior SVD}
\usage{
PCA2ACTIONred_full(S, x, sdev, rotation)
}
\arguments{
\item{S}{Input matrix ("sparseMatrix")}

\item{U}{Left singular vectors}

\item{s}{signular values}

\item{V}{Right singular vectors}
}
\value{
A named list with S_r, V, lambda, and exp_var. \itemize{
\item S_r: reduced kernel matrix of size reduced_dim x #samples.
\item V: Associated left singular-vectors (useful for reconstructing discriminative scores for features, such as genes).
\item lambda, exp_var: Summary statistics of the sigular-values.
}
}
\description{
Computes reduced kernel matrix for a given (single-cell) profile and prior SVD
}
\examples{
S = logcounts(sce)
irlba.out = irlba::prcomp_irlba(S, n = 50, retx = TRUE, center = T)
red.out = PCA2ACTIONred_full(S, irlba.out$x, irlba.out$rotation, as.matrix(irlba.out$sdev))
Sr = red.out$S_r
}
