% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.3D}
\alias{plot.ACTIONet.3D}
\title{Main ACTIONet 3D plotting functions}
\usage{
\method{plot}{ACTIONet.3D}(
  ace,
  labels = NULL,
  trans_attr = NULL,
  trans_th = -1,
  trans_fac = 1,
  point_size = 1,
  palette = CPal_default,
  coordinate_slot = "ACTIONet3D"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to be projected on the ACTIONet plot}

\item{trans_attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans_th, }{trans_fac Control the effect of transparency mapping}

\item{point_size}{Size of nodes in the ACTIONet plot}

\item{palette}{Color palette (named vector or a name for a given known palette)}

\item{coordinate_slot}{Entry in colMaps(ace) containing the plot coordinates (default:'ACTIONet2D')}

\item{title}{Main title of the plot}
}
\value{
Visualized ACTIONet
}
\description{
Main ACTIONet 3D plotting functions
}
\examples{
ace = run.ACTIONet(sce)
plot.ACTIONet.3D(ace, ace$assigned_archetype, trans_attr = ace$node_centrality)
}
