% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.feature.view}
\alias{plot.ACTIONet.feature.view}
\title{Plots projection of features onto the ACTIONet plot}
\usage{
\method{plot}{ACTIONet.feature.view}(
  ace,
  feature.enrichment.table,
  top.features = 5,
  CPal = NULL,
  title = "Feature view",
  label.text.size = 1,
  renormalize = F,
  footprint_slot = "H_unified"
)
}
\arguments{
\item{ace}{ACTIONet output}

\item{feature.enrichment.table}{An arbitrary enrichment table with columns being archetypes and rows being specificity of features}

\item{top.features}{Number of features to return}

\item{CPal}{Color palette to use}

\item{title}{Main title of the plot}

\item{reorder.columns}{Whether to optimally re-order columns of the enrichment table}
}
\value{
Featur view
}
\description{
Plots projection of features onto the ACTIONet plot
}
\examples{
feature.enrichment.table = as.matrix(rowMaps(ace)[['unified_feature_specificity']])
plot.ACTIONet.feature.view(ace, feature.enrichment.table, 5)
}
