% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ACTIONet.interactive}
\alias{plot.ACTIONet.interactive}
\title{Interactive ACTIONet visualizetion with Plotly}
\usage{
\method{plot}{ACTIONet.interactive}(
  ace,
  label_attr = NULL,
  trans_attr = NULL,
  trans_fac = 1.5,
  trans_th = -0.5,
  point_size = 3,
  stroke_size = point_size * 0.1,
  stroke_color = "black",
  palette = CPal_default,
  show_legend = TRUE,
  coordinate_attr = "ACTIONet2D",
  point_order = NULL,
  alt_hovertext = NULL,
  threeD = FALSE
)
}
\arguments{
\item{label_attr}{Character vector of length NROW(ace) or colname of 'colData(ace)' containing cell labels of interest (clusters, cell types, etc.).}

\item{trans_attr}{Numeric vector of length NROW(ace) or colname of 'colData(ace)' used to compute point transparency. Smaller values are more transparent.}

\item{trans_fac}{Transparency modifier (default:1.5).}

\item{trans_th}{Minimum Z-score for which points with 'scale(trans_attr) < trans_th' are masked  (default:-0.5).}

\item{point_size}{Size of points in ggplot (default:3).}

\item{stroke_size}{Size of points outline (stroke) in ggplot (default:point_size*0.1).}

\item{stroke_color}{Marker stroke color. Can be any (single) color value accepted by plotly.}

\item{palette}{Color palette. character vector of hex colors or a palette name to pass to 'ggpubr::get_palette()').}

\item{show_legend}{Show legend based on labels. Ignored if 'label_attr=NULL' (default:'TRUE').}

\item{coordinate_attr}{Name of entry in colMaps(ace) containing the 2D or 3D plot coordinates if 'data' is an 'ACTIONetExperiment' object (default:'ACTIONet2D' or 'ACTIONet3D' if 'threeD=TRUE').}

\item{point_order}{Integer vector specifying order in which to plot individual points (default:'NULL').}

\item{alt_hovertext}{Vector of values to display on hover.  (default:'NULL').}

\item{threeD}{Visualize plot in 3D using 'scatter3D'. Required 3D coordinate input (defaults to 'ACTIONet3D' if data is 'ACTIONetExperiment') (default:'FALSE').}

\item{data}{'ACTIONetExperiment' object or numeric matrix of X-Y(-Z) coordinates.}
}
\value{
plotly object
}
\description{
Interactive ACTIONet visualizetion with Plotly
}
\examples{

plot.ACTIONet.interactive(ace, ace$assigned_archetype, threeD = TRUE)
}
