% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.top.k.features}
\alias{plot.top.k.features}
\title{Plots heatmap of the top-ranked features of an enrichment table}
\usage{
\method{plot}{top.k.features}(
  feature.enrichment.table,
  top.features = 3,
  normalize = T,
  reorder.columns = T,
  row.title = "Archetypes",
  column.title = "Genes",
  rowPal = "black"
)
}
\arguments{
\item{feature.enrichment.table}{An arbitrary enrichment table with columns being archetypes and rows being specificity of features}

\item{top.features}{Number of features to return}

\item{reorder.columns}{Whether to optimally re-order columns of the enrichment table}
}
\value{
Enrichment heatmap
}
\description{
Plots heatmap of the top-ranked features of an enrichment table
}
\examples{
feature.enrichment.table = as.matrix(rowFactors(ace)[["H_unified_upper_significance"]])
plot.top.k.features(feature.enrichment.table, 3)
}
