% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{prune.scores.using.ACTIONet}
\alias{prune.scores.using.ACTIONet}
\title{Uses sweepcup algorithm on the ACTIONet to prune/filter a given score per node.}
\usage{
prune.scores.using.ACTIONet(ACTIONet.out, scores)
}
\arguments{
\item{ACTIONet.out}{Input results to be clustered
(alternatively it can be the ACTIONet igraph object)}

\item{scores}{A vector with one value per ACTIONet node}
}
\value{
Pruned scores
}
\description{
Uses sweepcup algorithm on the ACTIONet to prune/filter a given score per node.
}
\examples{
pruned.scores = prune.scores.using.ACTIONet(ACTIONet.out, scores)
}
