% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prune_archetypes}
\alias{prune_archetypes}
\title{Filters multi-level archetypes and concatenate filtered archetypes.
(Pre-ACTIONet archetype processing)}
\usage{
prune_archetypes(
  C_trace,
  H_trace,
  min_specificity_z_threshold = -1,
  min_cells = 3L
)
}
\arguments{
\item{C_trace, H_trace}{Output of ACTION}

\item{min_specificity_z_threshold}{Defines the stringency of pruning nonspecific archetypes.
The larger the value, the more archetypes will be filtered out (default=-1)}
}
\value{
A named list: \itemize{
\item selected_archs: List of final archetypes that passed the filtering/pruning step.
\item C_stacked,H_stacked: Horizontal/Vertical concatenation of filtered C and H matrices, respectively.
}
}
\description{
Filters multi-level archetypes and concatenate filtered archetypes.
(Pre-ACTIONet archetype processing)
}
\examples{
S = logcounts(sce)
reduction.out = reduce(S, reduced_dim = 50)
S_r = reduction.out$S_r
ACTION.out = run_ACTION(S_r, k_max = 10)
reconstruction.out = reconstruct_archetypes(S, ACTION.out$C, ACTION.out$H)
}
