% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reconstructACTIONet}
\alias{reconstructACTIONet}
\title{Reconstructs the ACTIONet graph with the new parameters (uses prior decomposition)}
\usage{
reconstructACTIONet(
  ace,
  network_density = 1,
  distance_metric = "jsd",
  nn_approach = "k*nn",
  mutual_edges_only = TRUE,
  layout_compactness = 50,
  layout_epochs = 1000,
  layout_algorithm = c("tumap", "umap"),
  layout_in_parallel = TRUE,
  thread_no = 0,
  reduction_slot = "ACTION",
  new_net_slot = "ACTIONet",
  H_stacked_slot = "H_stacked",
  seed = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object.}

\item{network_density}{Density factor of ACTIONet graph. (default=1)}

\item{distance_metric}{Distance metric with which to compute cell-to-cell similarity during network construction. Options are 'jsd' (Jensen-Shannon divergence), L2-norm ('l2'), and inner product ('ip'). (default='jsd')}

\item{nn_approach}{Nearest-neighbor algorithm to use for network construction. Options are k-nearest neighbors ('knn') and k*-nearest neighbors ('k*nn'). (default='k*nn')}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors. (default=TRUE)}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50).}

\item{layout_epochs}{Number of epochs for SGD algorithm. (default=1000)}

\item{layout_algorithm}{Algorithm for computing plot layout. t-UMAP ("tumap") or UMAP ("umap"). Not case sensitive. (default="tumap")}

\item{layout_in_parallel}{Run layout construction using multiple cores. May result in marginally different outputs across runs due to parallelization-induced randomization. (default=TRUE)}

\item{thread_no}{Number of parallel threads. (default=0)}

\item{reduction_slot}{Slot in colMaps(ace) containing reduced kernel. (default='ACTION')}

\item{new_net_slot}{Name of slot in colMaps(ace) to store ACTIONet adjacency matrix. (default='ACTIONet')}

\item{seed}{Seed for random initialization. (default=0)}
}
\value{
ace Updated ace object
}
\description{
Reconstructs the ACTIONet graph with the new parameters (uses prior decomposition)
}
\examples{
plot.ACTIONet(ace)
ace.updated <- reconstructACTIONet(ace, network_density = 0.1)
plot.ACTIONet(ace.updated)
}
