% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduce.sce}
\alias{reduce.sce}
\title{Takes a `SingleCellExperiment` object and adds the reduced kernel matrix}
\usage{
reduce.sce(sce, reduced_dim = 50, max.iter = 5,
  data.slot = "logcounts", normalization.method = "default",
  reduction.slot = "ACTION", seed = 0, SVD_algorithm = 0)
}
\arguments{
\item{sce}{Input sce object}

\item{reduced_dim}{Dimension of SVD used for reducing kernel matrix}

\item{max.iter}{Number of SVD iterations}

\item{norm.method}{Normalization method to use. See normalize.sce() function (default:"default")
(used only if the sce object is not already normalized)}

\item{passphrase}{Passphrase for encrypting column names of the sce object for anonymization}
}
\value{
ACTIONetExperiment object (ACE), derived from SingleCellExperiment (SCE), with added ReducedDims(sce)[["S_r"]]
}
\description{
Takes a `SingleCellExperiment` object and adds the reduced kernel matrix
}
\examples{
sce = import.sce.from.10X(input_path)
sce = reduce.sce(sce)
}
