% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_kernel}
\alias{reduce_kernel}
\title{Computes reduced kernel matrix for a given (single-cell) profile}
\usage{
reduce_kernel(
  S,
  reduced_dim = 50L,
  iter = 5L,
  seed = 0L,
  reduction_algorithm = 0L,
  SVD_algorithm = 0L,
  prenormalize = FALSE
)
}
\arguments{
\item{S}{Input matrix ("sparseMatrix")}

\item{reduced_dim}{Dimension of the reduced kernel matrix (default=50)}

\item{seed}{Random seed (default=0)}

\item{reduction_algorithm}{Kernel reduction algorithm. Currently only ACTION method (1) is implemented (default=1)}

\item{SVD_algorithm}{SVD algorithm to use. Currently supported methods are Halko (1) and Feng (2) (default=1)}

\item{iters}{Number of SVD iterations (default=5)}
}
\value{
A named list with S_r, V, lambda, and exp_var. \itemize{
\item S_r: reduced kernel matrix of size reduced_dim x #samples.
\item V: Associated left singular-vectors (useful for reconstructing discriminative scores for features, such as genes).
\item lambda, exp_var: Summary statistics of the sigular-values.
}
}
\description{
Computes reduced kernel matrix for a given (single-cell) profile
}
\examples{
S = logcounts(sce)
reduction.out = reduce(S, reduced_dim = 50)
S_r = reduction.out$S_r
}
