% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{reducedDimNames}
\alias{reducedDimNames}
\title{Get names(colMaps()) as reducedDimNames()}
\usage{
reducedDimNames(x, ...)
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get names(colMaps()) as reducedDimNames()
}
