% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{reducedDims}
\alias{reducedDims}
\title{Get colMaps() as reducedDims()}
\usage{
reducedDims(x, ...)

reducedDims(x, ...)
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}

\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get colMaps() as reducedDims()

Get colMaps() as reducedDims()
}
