% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove.cells}
\alias{remove.cells}
\title{Remove cells from all the entries in the ACTIONet output list}
\usage{
remove.cells(ACTIONet.out, filtered.cells, force = TRUE)
}
\arguments{
\item{ACTIONet.out}{Output list to be pruned}

\item{filtered.cells}{Cells to remove}

\item{force}{Whether to allow to remove influential cells of each archetype (default=TRUE)}
}
\value{
Updated ACTIONet.out
}
\description{
Remove cells from all the entries in the ACTIONet output list
}
\examples{
sce = reduce(sce)
ACTIONet.out = run_ACTIONet(sce)
ACTIONet.out.pruned = remove.cells(ACTIONet.out, 1:100)
}
