% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rerun.layout}
\alias{rerun.layout}
\title{Rerun layout on the ACTIONet graph with new parameters}
\usage{
rerun.layout(
  ace,
  layout_compactness = 50,
  layout_epochs = 1000,
  layout_algorithm = 0,
  network_density = 1,
  mutual_edges_only = TRUE,
  thread_no = 0,
  reduction_slot = "ACTION",
  net_slot = "ACTIONet",
  seed = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object.}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50).}

\item{layout_epochs}{Number of epochs for SGD algorithm (default=1000).}

\item{layout_algorithm}{Algorithm for computing plot layout. Set to 0 for TUMAP, or 1 for UMAP (default=0).}

\item{network_density}{Density factor of ACTIONet graph (default=1).}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors (default=TRUE).}

\item{thread_no}{Number of parallel threads (default=0).}

\item{reduction_slot}{Slot in colMaps(ace) containing reduced kernel (default='ACTION').}

\item{net_slot}{Slot in colMaps(ace) containing ACTIONet adjacency matrix (default='ACTIONet').}

\item{seed}{Seed for random initialization (default=0).}
}
\value{
ace Updated ace object
}
\description{
Rerun layout on the ACTIONet graph with new parameters
}
\examples{
plot.ACTIONet(ace)
ace.updated = rerun.layout(ace, layout_compactness = 20)
plot.ACTIONet(ace.updated)
}
