% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowFactors<-}
\alias{rowFactors<-}
\title{Set rowFactors}
\usage{
rowFactors(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set rowFactors
}
