% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_ace.R, R/set_ace.R
\name{rowFactors}
\alias{rowFactors}
\alias{rowFactors,ACTIONetExperiment-method}
\alias{rowFactors<-,ACTIONetExperiment-method}
\title{Get rowFactors}
\usage{
rowFactors(x, ...)

\S4method{rowFactors}{ACTIONetExperiment}(x, withDimnames = TRUE)

\S4method{rowFactors}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices

List of matrices
}
\description{
Get rowFactors

Get row-associated factors

Set row-associated factors
}
