% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowMapMeta}
\alias{rowMapMeta}
\title{Get rowMapMeta}
\usage{
rowMapMeta(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get rowMapMeta
}
