% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get_ace.R, R/set_ace.R
\docType{methods}
\name{rowMaps}
\alias{rowMaps}
\alias{rowMaps,ACTIONetExperiment-method}
\alias{rowMaps<-,ACTIONetExperiment-method}
\title{Get rowMaps}
\usage{
rowMaps(x, ...)

\S4method{rowMaps}{ACTIONetExperiment}(x, all = T)

\S4method{rowMaps}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices

List of matrices
}
\description{
Get rowMaps

Get row-associated factors

Set row-associated factors
}
