% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowReductions<-}
\alias{rowReductions<-}
\title{Set row reductions}
\usage{
rowReductions(x, ...) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set row reductions
}
