% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run.ACTIONet}
\alias{run.ACTIONet}
\title{A wrapper function to call all main functions of the ACTIONet}
\usage{
run.ACTIONet(
  sce,
  k_max = 30,
  min_specificity_z_threshold = -1,
  network_density = 1,
  mutual_edges_only = TRUE,
  layout_compactness = 50,
  n_epochs = 500,
  thread_no = 4,
  sce.data.attr = "logcounts",
  reduction_slot = "S_r"
)
}
\arguments{
\item{sce}{Reduced `SingleCellExperiment` object (output of reduce.sce() function).}

\item{k_max}{Maximum depth of decompositions (default=30).}

\item{min_specificity_z_threshold}{Defines the stringency of pruning nonspecific archetypes. 
The larger the value, the more archetypes will be filtered out (default=-1).}

\item{network_density}{Density factor of ACTIONet graph (default=1).}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors (default=TRUE).}

\item{n_epochs}{Number of epochs for SGD algorithm (default=500).}

\item{thread_no}{Number of parallel threads (default=4)}

\item{sce.data.attr}{Corresponding slot in the `sce` object the normalized counts (default="logcounts")}

\item{reduction_slot}{Slot in the ReducedDims(sce) that holds reduced kernel (default="S_r")}

\item{compactness_level}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50)}
}
\value{
A named list: \itemize{
\item ACTION.out: Output of run_ACTION() with C and H traces.
\item pruning.out: Output of prune_archetypes() with C_stacked and H_stacked inside.
\item ACTIONet: ACTIONet graph, constructed by build_ACTIONet(), in the igraph format
\item vis.out: Output of the layout_ACTIONet() function. It includes coordinates and colors for ACTIONet nodes.
\item unification.out: Output of the unify_archetypes() function. Includes C_unified, H_unified, and sample_assignments.
}
}
\description{
A wrapper function to call all main functions of the ACTIONet
}
\examples{
sce = reduce(sce)
ACTIONet.out = run_ACTIONet(sce)
}
