% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_AA}
\alias{run_AA}
\title{Runs basic archetypal analysis}
\usage{
run_AA(A, W0, max_it = 50L, min_delta = 1e-16)
}
\arguments{
\item{A}{Inpu matrix}

\item{W0}{Starting archetypes}

\item{max_it, min_delta}{Convergence parameters for archetypal analysis}
}
\value{
A named list with entries 'C' and 'H', each a list for different values of k
}
\description{
Runs basic archetypal analysis
}
\examples{
S_r = t(reducedDims(ace)$ACTION)
SPA.out = run_SPA(S_r, 10)
W0 = S_r[, SPA.out$selected_columns]
AA.out = run_AA(S_r, W0)
H = AA.out$H
cell.assignments = apply(H, 2, which.max)
}
