% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_ACTION_plus}
\alias{run_ACTION_plus}
\title{Runs multi-level ACTION decomposition method}
\usage{
run_ACTION_plus(S_r, k_min = 2L, k_max = 30L, max_it = 50L,
  min_delta = 1e-16, max_trial = 3L)
}
\arguments{
\item{S_r}{Reduced kernel matrix}

\item{k_min}{Minimum number of archetypes to consider (default=2)}

\item{k_max}{Maximum number of archetypes to consider, or "depth" of decomposition (default=30)}

\item{max_it, min_delta}{Convergence parameters for archetypal analysis}

\item{thread_no}{Number of parallel threads (default=4)}
}
\value{
A named list with entries 'C' and 'H', each a list for different values of k
}
\description{
Runs multi-level ACTION decomposition method
}
\examples{
ACTION.out = run_ACTION(S_r, k_max = 10)
H8 = ACTION.out$H[[8]]
cell.assignments = apply(H8, 2, which.max)
}
