% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_online_ACTION}
\alias{run_online_ACTION}
\title{Runs multi-level Online ACTION decomposition method (under development)}
\usage{
run_online_ACTION(S_r, samples, k_min = 2L, k_max = 30L,
  thread_no = 0L)
}
\arguments{
\item{S_r}{Reduced kernel matrix}

\item{samples}{List of sampled cells to use for updating archetype decomposition}

\item{k_min}{Minimum number of archetypes to consider (default=2)}

\item{k_max}{Maximum number of archetypes to consider, or 'depth' of decomposition (default=30)}

\item{thread_no}{Number of parallel threads (default = 0)}
}
\value{
A named list with entries 'C' and 'H', each a list for different values of k
}
\description{
Runs multi-level Online ACTION decomposition method (under development)
}
\examples{
ACTION.out = run_online_ACTION(S_r, k_max = 10)
}
