% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_weighted_ACTION}
\alias{run_weighted_ACTION}
\title{Runs multi-level ACTION decomposition method}
\usage{
run_weighted_ACTION(
  S_r,
  w,
  k_min = 2L,
  k_max = 30L,
  thread_no = 8L,
  max_it = 50L,
  min_delta = 0.01
)
}
\arguments{
\item{S_r}{Reduced kernel matrix}

\item{w}{Weight vector for each observation}

\item{k_min}{Minimum number of archetypes to consider (default=2)}

\item{k_max}{Maximum number of archetypes to consider, or "depth" of decomposition (default=30)}

\item{thread_no}{Number of parallel threads (default=4)}
}
\value{
A named list with entries 'C' and 'H', each a list for different values of k
}
\description{
Runs multi-level ACTION decomposition method
}
\examples{
ACTION.out = run_ACTION(S_r, k_max = 10)
H8 = ACTION.out$H[[8]]
cell.assignments = apply(H8, 2, which.max)
}
