% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runACTIONet}
\alias{runACTIONet}
\title{Run main ACTIONet pipeline}
\usage{
runACTIONet(
  ace,
  batch = NULL,
  k_min = 2,
  k_max = 30,
  assay_name = "logcounts",
  batch_attr = NULL,
  batch_correction_method = "Harmony",
  rerun_reduction = FALSE,
  reduced_dim = 50,
  reduction_slot = "ACTION",
  net_slot_out = "ACTIONet",
  min_cells_per_arch = 2,
  max_iter_ACTION = 50,
  min_specificity_z_thresh = -3,
  distance_metric = "jsd",
  nn_approach = "k*nn",
  network_density = 1,
  mutual_edges_only = TRUE,
  imputation_alpha = 0.9,
  layout_compactness = 50,
  layout_epochs = 1000,
  layout_algorithm = "TUMAP",
  layout_in_parallel = TRUE,
  unification_violation_threshold = 0,
  footprint_alpha = 0.15,
  thread_no = 0,
  full_trace = FALSE,
  seed = 0
)
}
\arguments{
\item{ace}{`ACTIONetExperiment` containing an appropriate reduction in 'reduction_slot'.}

\item{k_min}{Minimum depth of decompositions. (default=2)}

\item{k_max}{Maximum depth of decompositions. (default=30)}

\item{assay_name}{Name of assay to be used. (default='logcounts')}

\item{batch_attr}{Which attribute to use for batch correction. No batch correction is done if it is NULL (default=NULL)}

\item{batch_correction_method}{Batch-correction method. Currently can be MNN, Harmony, and BatchOrtho (default="Harmony")}

\item{rerun_reduction}{Which attribute to use for batch correction. No batch correction is done if it is NULL (default=TRUE)}

\item{reduced_dim}{Which attribute to use for batch correction. No batch correction is done if it is NULL (default=50)}

\item{reduction_slot}{Slot in colMaps(ace) containing reduced kernel. (default='ACTION')}

\item{net_slot_out}{Name of slot in colMaps(ace) to store ACTIONet adjacency matrix. (default='ACTIONet')}

\item{min_cells_per_arch}{Minimum number of observations required to construct an archetype. (default=2)}

\item{max_iter_ACTION}{Maximum number of iterations for ACTION algorithm. (default=50)}

\item{min_specificity_z_thresh}{Defines the stringency of pruning nonspecific archetypes.
The larger the value, the more archetypes will be filtered out. (default=-3)}

\item{distance_metric}{Distance metric with which to compute cell-to-cell similarity during network construction. Options are 'jsd' (Jensen-Shannon divergence), L2-norm ('l2'), and inner product ('ip'). (default='jsd')}

\item{nn_approach}{Nearest-neighbor algorithm to use for network construction. Options are k-nearest neighbors ('knn') and k*-nearest neighbors ('k*nn'). (default='k*nn')}

\item{network_density}{Density factor of ACTIONet graph. (default=1)}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors. (default=TRUE)}

\item{imputation_alpha}{Network diffusion parameter to smooth PCs (S_r). (default=0.9)}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout. (default=50)}

\item{layout_epochs}{Number of epochs for SGD algorithm. (default=1000)}

\item{layout_algorithm}{Algorithm for computing plot layout. can be "UMAP", "TUMAP" (faster), or "forced_atlas". (default="TUMAP")}

\item{layout_in_parallel}{Run layout construction using multiple cores. May result in marginally different outputs across runs due to parallelization-induced randomization. (default=TRUE)}

\item{unification_violation_threshold}{Archetype unification resolution parameter. (default=0)}

\item{footprint_alpha}{Archetype smoothing parameter. (default=0.85)}

\item{thread_no}{Number of parallel threads. (default=0)}

\item{full_trace}{Return list of all intermediate output. Intended for debugging. (default='FALSE')}

\item{seed}{Seed for random initialization. (default=0)}
}
\value{
\itemize{
\item If full_trace='FALSE'(default): ACTIONetExperiment object.
\item If full_trace='TRUE': Named list containing an ACTIONetExperiment object and a log of ACTIONet function calls.
}
}
\description{
Run main ACTIONet pipeline
}
\examples{
ace <- runACTIONet(ace)
}
