% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{select.top.k.features}
\alias{select.top.k.features}
\title{Report the top-rank features from a given enrichment table}
\usage{
\method{select}{top.k.features}(
  feature.enrichment.table,
  top.features = 3,
  normalize = F,
  reorder.columns = T
)
}
\arguments{
\item{top.features}{Number of features to return}

\item{reorder.columns}{Whether to optimally re-order columns of the enrichment table}
}
\value{
Sorted table with the selected top-ranked
}
\description{
Report the top-rank features from a given enrichment table
}
\examples{
feature.enrichment.table = as.matrix(rowFactors(ace)[["H_unified_upper_significance"]])
enrichment.table.top = select.top.k.features(feature.enrichment.table, 3)
}
