% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVD2ACTIONred_full}
\alias{SVD2ACTIONred_full}
\title{Computes reduced kernel matrix for a given (single-cell) profile and prior SVD}
\usage{
SVD2ACTIONred_full(S, u, d, v)
}
\arguments{
\item{S}{Input matrix ("sparseMatrix")}

\item{U}{Left singular vectors}

\item{s}{signular values}

\item{V}{Right singular vectors}
}
\value{
A named list with S_r, V, lambda, and exp_var. \itemize{
\item S_r: reduced kernel matrix of size reduced_dim x #samples.
\item V: Associated left singular-vectors (useful for reconstructing discriminative scores for features, such as genes).
\item lambda, exp_var: Summary statistics of the sigular-values.
}
}
\description{
Computes reduced kernel matrix for a given (single-cell) profile and prior SVD
}
\examples{
S = logcounts(sce)
irlba.out = irlba::irlba(S, nv = 50)
red.out = SVD2ACTIONred_full(S, irlba.out$u, as.matrix(irlba.out$d), irlba.out$v)
Sr = red.out$S_r
}
