% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unify_archetypes}
\alias{unify_archetypes}
\title{Identifies and aggregates redundant archetypes into equivalent classes
(Post-ACTIONet archetype processing)}
\usage{
unify_archetypes(S_r, C_stacked, H_stacked, min_overlap = 10, resolution = 1)
}
\arguments{
\item{S_r}{Reduced kernel profile}

\item{C_stacked, H_stacked}{Output of reconstruct_archetypes()}

\item{G}{Adjacency matrix of the ACTIONet graph}

\item{archetypes}{Archetype profile (S*C)}

\item{minPoints, }{minClusterSize, outlier_threshold HDBSCAN parameters}

\item{reduced_dim}{Kernel reduction}
}
\value{
A named list: \itemize{
\item archetype_groups: Equivalent classes of archetypes (non-redundant)
\item C_unified,H_unified: C and H matrices of unified archetypes
\item sample_assignments: Assignment of samples/cells to unified archetypes
}
}
\description{
Identifies and aggregates redundant archetypes into equivalent classes
(Post-ACTIONet archetype processing)
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
G = build_ACTIONet(prune.out$H_stacked)
unification.out = unify_archetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
}
