% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unsigned_cluster}
\alias{unsigned_cluster}
\title{Computes graph clustering using Leiden algorith over unsigned graphs}
\usage{
unsigned_cluster(A, resolution_parameter = 1, initial_clusters_ = NULL,
  seed = 0L)
}
\arguments{
\item{resolution_parameter}{Granularity of clustering. Larger values result in more clusters (default = 1.0)}

\item{initial_clusters_}{Initialization vector for clusters (if available)}

\item{seed}{Random seed}

\item{G}{Adjacency matrix of the input graph}
}
\value{
clusters Assignment vector of samples to clusters
}
\description{
Computes graph clustering using Leiden algorith over unsigned graphs
}
\examples{
clusters = unsigned_cluster(G)
}
