% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ViperRegulons}
\alias{ViperRegulons}
\title{Regulons for Viper analysis}
\format{
:
  A list of regulons as input to the Viper method
}
\source{
\url{https://www.bioconductor.org/packages/release/bioc/html/viper.html}
}
\usage{
ViperRegulons
}
\description{
Regulons for Viper analysis
}
\keyword{datasets}
