% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize.markers}
\alias{visualize.markers}
\title{Projects a set of markers on the ACTIONet
It also optionally imputes the markers.}
\usage{
visualize.markers(
  ace,
  markers,
  features_use = NULL,
  alpha_val = 0.9,
  assay_name = "logcounts",
  trans_attr = NULL,
  trans_th = -0.5,
  trans_fac = 3,
  grad_palette = "magma",
  point_size = 1,
  net_attr = "ACTIONet",
  coordinate_attr = "ACTIONet2D",
  single_plot = FALSE
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{markers}{Set of row features (e.g. genes) to visualize .}

\item{alpha_val}{Between [0, 1]. If it is greater than 0, smoothing of scores would be performed}

\item{trans_attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans_th, }{trans_fac Control the effect of transparency mapping}

\item{grad_palette}{Color palette (named vector or a name for a given known palette)}

\item{point_size}{Size of nodes in the ACTIONet plot}

\item{coordinate_slot}{Entry in colMaps(ace) containing the plot coordinates (default:'ACTIONet2D')}
}
\value{
Visualized ACTIONet with projected scores
}
\description{
Projects a set of markers on the ACTIONet
It also optionally imputes the markers.
}
\examples{
ace <- run.ACTIONet(sce)
visualize.markers(ace, markers = c("CD14", "CD19", "CD3G"), trans_attr = ace$node_centrality)
}
