% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{cluster.ACTIONet}
\alias{cluster.ACTIONet}
\title{A wrapper function For Leiden algorithm applied to an ACE object}
\usage{
cluster.ACTIONet(
  ace,
  resolution_parameter = 1,
  net.slot = "ACTIONet",
  init.slot = "assigned_archetype",
  seed = 0
)
}
\arguments{
\item{ace}{Input results to be clustered}

\item{resolution_parameter}{Resolution of the clustering.
The higher the resolution, the more clusters we will get (default=0.5).}

\item{seed}{Random seed}

\item{arch.init}{Whether to use archetype-assignments to initialize clustering (default=TRUE)}
}
\value{
ace with added annotation
}
\description{
A wrapper function For Leiden algorithm applied to an ACE object
}
\examples{
clusters = cluster.ACTIONet(ace)
plot.ACTIONet(ace, clusters)
}
