% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{clusterFeatureSpecificity}
\alias{clusterFeatureSpecificity}
\title{Computes feature specificity scores for each cluster.}
\usage{
clusterFeatureSpecificity(
  obj,
  cluster_attr,
  output_prefix = "cluster_spec",
  assay_name = "logcounts",
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Computes feature specificity scores for each cluster.
}
