% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colEmbeddings}
\alias{colEmbeddings}
\title{Get column embeddings}
\usage{
colEmbeddings(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get column embeddings
}
