% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colMapMeta}
\alias{colMapMeta}
\title{Get colMapTypes}
\usage{
colMapMeta(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get colMapTypes
}
