% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colReductions<-}
\alias{colReductions<-}
\title{Set column reductions}
\usage{
colReductions(obj, ...) <- value
}
\arguments{
\item{obj}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Set column reductions
}
