% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_ace.R
\docType{methods}
\name{colReductions}
\alias{colReductions}
\alias{colReductions<-,ACTIONetExperiment-method}
\title{Get column reductions}
\usage{
colReductions(x, ...)

\S4method{colReductions}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices
}
\description{
Get column reductions

Set column-associated reductions
}
