% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_ace.R
\name{colReductions}
\alias{colReductions}
\alias{colReductions<-,ACTIONetExperiment-method}
\title{Get column reductions}
\usage{
colReductions(obj, ...)

\S4method{colReductions}{ACTIONetExperiment}(obj, val) <- value
}
\arguments{
\item{obj}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices
}
\description{
Get column reductions

Set column-associated reductions
}
