% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_cluster_feature_specificity_full}
\alias{compute_cluster_feature_specificity_full}
\title{Compute feature specificity (from cluster assignments)}
\usage{
compute_cluster_feature_specificity_full(S, sample_assignments)
}
\arguments{
\item{S}{Input matrix ("matrix")}

\item{sample_assignments}{Vector of cluster assignments}
}
\value{
A list with the over/under-logPvals
}
\description{
Compute feature specificity (from cluster assignments)
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
G = build_ACTIONet(prune.out$H_stacked)
unification.out = unify_archetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
S.norm = renormalize_input_matrix(S, cell.clusters)
logPvals.list = compute_cluster_feature_specificity(S.norm, cell.clusters)
specificity.scores = logPvals.list$upper_significance
}
