% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{correct.cell.annotations}
\alias{correct.cell.annotations}
\title{Uses a variant of the label propagation algorithm to correct likely noisy labels}
\usage{
correct.cell.annotations(
  ACTIONet.out,
  annotation.in,
  annotation.out,
  LFR.threshold = 2,
  double.stochastic = FALSE,
  max_iter = 3,
  adjust.levels = T,
  min.cell.fraction = 0.001
)
}
\arguments{
\item{ACTIONet.out}{Input results to be clustered
(alternatively it can be the ACTIONet igraph object)}

\item{annotation.in}{Annotations to correct with missing values (NA) in it.
It can be either a named annotation (inside ACTIONet.out$annotations) or a label vector.}

\item{annotation.out}{Name of the updated annotations to be stored.}

\item{LFR.threshold}{How aggressively to update labels. The smaller the value, the more labels will be changed (default=2)}

\item{double.stochastic}{Whether to densify adjacency matrix before running label propagation (default=FALSE).}

\item{max_iter}{How many iterative rounds of correction/inference should be performed (default=3)}

\item{min.cell.fraction}{Annotations with less that this fraction will be removed}
}
\value{
ACTIONet.out with updated annotations added to ACTIONet.out$annotations
}
\description{
Uses a variant of the label propagation algorithm to correct likely noisy labels
}
\examples{
ACTIONet.out = add.cell.annotations(ACTIONet.out, cell.labels, "input_annotations")
ACTIONet.out = correct.cell.annotations(ACTIONet.out, "input_annotations", "updated_annotations")
}
