% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decomp.ACTION}
\alias{decomp.ACTION}
\title{Run ACTION_decomposition}
\usage{
decomp.ACTION(
  X,
  k,
  max_iter = 50,
  min_delta = 1e-300,
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Run ACTION_decomposition
}
