% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal_main.R
\name{.run.pruneArchetypes}
\alias{.run.pruneArchetypes}
\title{Prune nonspecific and/or unreliable archetypes}
\usage{
.run.pruneArchetypes(
  ace,
  C_trace,
  H_trace,
  specificity_th = -3,
  min_cells_per_arch = 2
)
}
\description{
Prune nonspecific and/or unreliable archetypes
}
