% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal_main.R
\name{.run.unifyArchetypes}
\alias{.run.unifyArchetypes}
\title{Identiy equivalent classes of archetypes and group them together}
\usage{
.run.unifyArchetypes(
  ace,
  unification_backbone_density = 0.5,
  unification_resolution = 1,
  unification_min_cluster_size = 3,
  reduction_slot = "ACTION",
  C_stacked_slot = "C_stacked",
  H_stacked_slot = "H_stacked",
  unified_suffix = "unified",
  footprint_slot_name = "assigned_archetype",
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Identiy equivalent classes of archetypes and group them together
}
