% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_tools.R
\name{infer.missing.cell.annotations}
\alias{infer.missing.cell.annotations}
\title{Uses a variant of the label propagation algorithm to infer missing labels}
\usage{
infer.missing.cell.annotations(ace, annotation.in, annotation.out,
  double.stochastic = FALSE, max_iter = 3, adjust.levels = T)
}
\arguments{
\item{ace}{Input results to be clustered
(alternatively it can be the ACTIONet igraph object)}

\item{annotation.in}{Annotations to correct with missing values (NA) in it.
It can be either a named annotation (inside ace$annotations) or a label vector.}

\item{annotation.out}{Name of the updated annotations to be stored.}

\item{double.stochastic}{Whether to densify adjacency matrix before running label propagation (default=FALSE).}

\item{max_iter}{How many iterative rounds of correction/inference should be performed (default=3)}

\item{adjust.levels}{Whether or not re-adjust labels at the end}
}
\value{
ace with updated annotations added to ace$annotations
}
\description{
Uses a variant of the label propagation algorithm to infer missing labels
}
\examples{
ace = infer.missing.cell.annotations(ace, sce$assigned_archetypes, 'updated_archetype_annotations')
}
