% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot.ACTIONet.interactive}
\alias{plot.ACTIONet.interactive}
\title{Main ACTIONet interactive plotting function}
\usage{
\method{plot}{ACTIONet.interactive}(
  ace,
  labels = NULL,
  trans_attr = NULL,
  trans_th = -1,
  trans_fac = 1,
  point_size = 1,
  palette = CPal_default,
  top_features = 7,
  blacklist_pattern = "\\\\.|^RPL|^RPS|^MRP|^MT-|^MT|MALAT1|B2M|GAPDH",
  threeD = FALSE,
  title = "ACTIONet",
  coordinate_slot = "ACTIONet2D"
)
}
\arguments{
\item{ace}{ACTIONet output object}

\item{labels}{Annotation of interest (clusters, celltypes, etc.) to be projected on the ACTIONet plot}

\item{trans_attr}{Additional continuous attribute to project onto the transparency of nodes}

\item{trans_th, }{trans_fac Control the effect of transparency mapping}

\item{point_size}{Size of nodes in the ACTIONet plot}

\item{palette}{Color palette (named vector or a name for a given known palette)}

\item{top_features}{Number of features to show per cell}

\item{blacklist_pattern}{List of genes to filter-out}

\item{threeD}{Whether to show the plot in 3D}

\item{title}{Main title of the plot}

\item{enrichment.table}{To project the top-ranked features interactively.}
}
\value{
Visualized ACTIONet
}
\description{
Main ACTIONet interactive plotting function
}
\examples{
ace = run.ACTIONet(sce)
plot.ACTIONet.interactive(ace, ace$assigned_archetype)
}
