% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduce.ace}
\alias{reduce.ace}
\title{Takes a `ACTIONetExperiment` object and adds the reduced kernel matrix}
\usage{
reduce.ace(
  ace,
  reduced_dim = 50,
  max_iter = 5,
  data_slot = "logcounts",
  norm_method = "default",
  reduction_slot = "ACTION",
  seed = 0,
  SVD_algorithm = 0
)
}
\arguments{
\item{ace}{Input ace object}

\item{reduced_dim}{Dimension of SVD used for reducing kernel matrix}

\item{max_iter}{Number of SVD iterations}

\item{norm.method}{Normalization method to use. See normalize.ace() function (default:'default')
(used only if the ace object is not already normalized)}

\item{passphrase}{Passphrase for encrypting column names of the ace object for anonymization}
}
\value{
ACTIONetExperiment object (ACE), derived from SummarizedExperiment (ace), with added colMaps(ace)[['S_r']]
}
\description{
Takes a `ACTIONetExperiment` object and adds the reduced kernel matrix
}
\examples{
ace = import.ace.from.10X(input_path)
ace = reduce.ace(ace)
}
