% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduce.ace}
\alias{reduce.ace}
\title{Takes a `ACTIONetExperiment` object and adds the reduced kernel matrix}
\usage{
reduce.ace(
  ace,
  reduced_dim = 50,
  max.iter = 5,
  data.slot = "logcounts",
  normalization.method = "default",
  reduction.slot = "ACTION",
  seed = 0,
  SVD_algorithm = 1
)
}
\arguments{
\item{ace}{Input ace object}

\item{reduced_dim}{Dimension of SVD used for reducing kernel matrix}

\item{max.iter}{Number of SVD iterations}

\item{norm.method}{Normalization method to use. See normalize.ace() function (default:"default")
(used only if the ace object is not already normalized)}

\item{passphrase}{Passphrase for encrypting column names of the ace object for anonymization}
}
\value{
ACTIONetExperiment object (ACE), derived from RangedSummarizedExperiment (ace), with added ReducedDims(ace)[["S_r"]]
}
\description{
Takes a `ACTIONetExperiment` object and adds the reduced kernel matrix
}
\examples{
ace = import.ace.from.10X(input_path)
ace = reduce.ace(ace)
}
