% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rerun.layout}
\alias{rerun.layout}
\title{Rerun layout on the ACTIONet graph with new parameters}
\usage{
rerun.layout(
  ace,
  layout_compactness = 50,
  layout_epochs = 500,
  thread_no = 1,
  reduction_slot = "ACTION2D"
)
}
\arguments{
\item{ace}{ACTIONetExperiment object containing the results}

\item{thread_no}{Number of parallel threads (default=8)}

\item{reduction_slot}{Slot in the colMaps(ace) that holds reduced kernel (default='S_r')}

\item{compactness_level}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50)}

\item{n_epochs}{Number of epochs for SGD algorithm (default=500).}
}
\value{
ace Updated ace object
}
\description{
Rerun layout on the ACTIONet graph with new parameters
}
\examples{
plot.ACTIONet(ace)
ace.updated = rerun.layout(ace, layout_compactness = 20)
plot.ACTIONet(ace.updated)
}
