% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rerunLayout}
\alias{rerunLayout}
\title{Rerun layout on the ACTIONet graph with new parameters}
\usage{
rerunLayout(
  ace,
  algorithm = c("umap", "tumap"),
  initial_coordinates = NULL,
  spread = 1,
  min_dist = 1,
  gamma = 1,
  n_epochs = 100,
  thread_no = 0,
  net_slot = "ACTIONet",
  seed = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object.}

\item{thread_no}{Number of parallel threads (default=0).}

\item{net_slot}{Slot in colMaps(ace) containing ACTIONet adjacency matrix (default='ACTIONet').}

\item{seed}{Seed for random initialization (default=0).}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout (default=50).}

\item{layout_epochs}{Number of epochs for SGD algorithm (default=250).}

\item{layout_algorithm}{Algorithm for computing plot layout. Options are UMAP ("umap") or t-UMAP ("tumap"). (default="umap")}

\item{network_density}{Density factor of ACTIONet graph (default=1).}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors (default=TRUE).}

\item{reduction_slot}{Slot in colMaps(ace) containing reduced kernel (default='ACTION').}
}
\value{
ace Updated ace object
}
\description{
Rerun layout on the ACTIONet graph with new parameters
}
\examples{
plot.ACTIONet(ace)
ace.updated <- rerunLayout(ace, compactness = 20)
plot.ACTIONet(ace.updated)
}
