% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_ace.R
\name{rowMapMeta}
\alias{rowMapMeta}
\alias{rowMapMeta<-,ACTIONetExperiment-method}
\title{Get rowMapMeta}
\usage{
rowMapMeta(x, ...)

\S4method{rowMapMeta}{ACTIONetExperiment}(x) <- value
}
\arguments{
\item{x}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices
}
\description{
Get rowMapMeta

Set row-associated factor annotations
}
