% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowMapTypes}
\alias{rowMapTypes}
\title{Get rowMapTypes}
\usage{
rowMapTypes(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get rowMapTypes
}
