% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_ace.R
\name{rowReductions}
\alias{rowReductions}
\alias{rowReductions<-,ACTIONetExperiment-method}
\title{Get row reductions}
\usage{
rowReductions(obj, ...)

\S4method{rowReductions}{ACTIONetExperiment}(obj, val) <- value
}
\arguments{
\item{obj}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices
}
\description{
Get row reductions

Set row-associated reductions
}
