% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{rowReductions}
\alias{rowReductions}
\title{Get row reductions}
\usage{
rowReductions(object, ...)
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\description{
Get row reductions
}
