% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_weighted_ACTION}
\alias{run_weighted_ACTION}
\title{Runs multi-level weighted ACTION decomposition method (under development)}
\usage{
run_weighted_ACTION(
  S_r,
  w,
  k_min = 2L,
  k_max = 30L,
  thread_no = 0L,
  max_it = 50L,
  min_delta = 1e-16
)
}
\arguments{
\item{S_r}{Reduced kernel matrix}

\item{w}{Weight vector for each observation}

\item{k_min}{Minimum number of archetypes to consider (default=2)}

\item{k_max}{Maximum number of archetypes to consider, or "depth" of decomposition (default=30)}

\item{thread_no}{Number of parallel threads (default=0)}
}
\value{
A named list with entries 'C' and 'H', each a list for different values of k
}
\description{
Runs multi-level weighted ACTION decomposition method (under development)
}
\examples{
ACTION.out = run_weighted_ACTION(S_r, w, k_max = 20)
}
