% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runACTIONet}
\alias{runACTIONet}
\title{Run main ACTIONet pipeline}
\usage{
runACTIONet(
  ace,
  k_min = 2,
  k_max = 30,
  assay_name = NULL,
  reduction_slot = NULL,
  reduction_normalization = 1,
  net_slot_out = "ACTIONet",
  min_cells_per_arch = 2,
  max_iter_ACTION = 50,
  specificity_th = -3,
  network_metric = "jsd",
  network_algorithm = "k*nn",
  network_density = 1,
  mutual_edges_only = TRUE,
  imputation_alpha = 0.9,
  layout_spread = 1,
  layout_min_dist = 1,
  layout_gamma = 1,
  layout_epochs = 100,
  layout_algorithm = c("umap", "tumap"),
  layout_parallel = TRUE,
  unification_backbone_density = 0.5,
  unification_resolution = 1,
  unification_min_cluster_size = 3,
  footprint_alpha = 0.85,
  compute_specificity_parallel = FALSE,
  smoothPC = TRUE,
  thread_no = 0,
  backbone_density = 0.5,
  seed = 0
)
}
\arguments{
\item{ace}{`ACTIONetExperiment` containing an appropriate reduction in 'reduction_slot'.}

\item{k_min}{Minimum depth of decompositions. (default=2)}

\item{k_max}{Maximum depth of decompositions. (default=30)}

\item{assay_name}{Name of assay to be used. (default='logcounts')}

\item{reduction_slot}{Slot in colMaps(ace) containing reduced kernel. (default='ACTION')}

\item{net_slot_out}{Name of slot in colMaps(ace) to store ACTIONet adjacency matrix. (default='ACTIONet')}

\item{min_cells_per_arch}{Minimum number of observations required to construct an archetype. (default=2)}

\item{max_iter_ACTION}{Maximum number of iterations for ACTION algorithm. (default=50)}

\item{specificity_th}{Defines the stringency of pruning nonspecific archetypes.
The larger the value, the more archetypes will be filtered out. (default=-3)}

\item{network_metric}{Distance metric with which to compute cell-to-cell similarity during network construction. Options are 'jsd' (Jensen-Shannon divergence), L2-norm ('l2'), and inner product ('ip'). (default='jsd')}

\item{network_algorithm}{Algorithm to use for network construction. Options are k-nearest neighbors ('knn') and k*-nearest neighbors ('k*nn'). (default='k*nn')}

\item{network_density}{Density factor of ACTIONet graph. (default=1)}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors. (default=TRUE)}

\item{imputation_alpha}{Network diffusion parameter to smooth PCs (S_r). (default=0.9)}

\item{layout_epochs}{Number of epochs for SGD algorithm. (default=250)}

\item{layout_algorithm}{Algorithm for computing plot layout. Options are UMAP ("umap") or t-UMAP ("tumap"). (default="umap")}

\item{layout_parallel}{Run layout construction using multiple cores. May result in marginally different outputs across runs due to parallelization-induced randomization. (default=TRUE)}

\item{footprint_alpha}{Archetype smoothing parameter. (default=0.85)}

\item{compute_specificity_parallel}{Run feature specificity enrichment using multiple cores. Setting this to `TRUE` on large datasets may cause an out of memory crash. (default=FALSE)}

\item{smoothPC}{Whether to smooth output of 'reduce.ace()'. Performs diffusion smoothing on the components of the reduction. Skipped if 'TRUE' and output of 'reduce.ace()' is missing from object. (default=TRUE)}

\item{thread_no}{Number of parallel threads. (default=0)}

\item{seed}{Seed for random initialization. (default=0)}

\item{layout_compactness}{A value between 0-100, indicating the compactness of ACTIONet layout. (default=50)}

\item{unification_th}{Archetype unification resolution parameter. (default=0)}
}
\value{
An ACTIONetExperiment object
}
\description{
Run main ACTIONet pipeline
}
\examples{
ace <- runACTIONet(ace)
}
