% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unify_archetypes}
\alias{unify_archetypes}
\title{Identifies and aggregates redundant archetypes into equivalent classes
(Post-ACTIONet archetype processing)}
\usage{
unify_archetypes(
  S_r,
  C_stacked,
  H_stacked,
  min_edge_weight = 0.5,
  min_coreness = 0L,
  resolution = 1,
  min_repeat = 2L,
  thread_no = 0L,
  alpha = 0.05,
  beta = 0.5,
  outlier_threshold = 0.5,
  minPoints = 5L,
  minClusterSize = 5L,
  cond_threshold = 10,
  normalization_type = 3L,
  preprocess_adj = FALSE,
  reduce_G = FALSE,
  method_type = 0L,
  sensitivity = 0.95
)
}
\arguments{
\item{S_r}{Reduced kernel profile}

\item{C_stacked, H_stacked}{Output of reconstruct_archetypes()}

\item{minPoints, }{minClusterSize, outlier_threshold HDBSCAN parameters}

\item{G}{Adjacency matrix of the ACTIONet graph}

\item{archetypes}{Archetype profile (S*C)}

\item{reduced_dim}{Kernel reduction}
}
\value{
A named list: \itemize{
\item archetype_groups: Equivalent classes of archetypes (non-redundant)
\item C_unified,H_unified: C and H matrices of unified archetypes
\item sample_assignments: Assignment of samples/cells to unified archetypes
}
}
\description{
Identifies and aggregates redundant archetypes into equivalent classes
(Post-ACTIONet archetype processing)
}
\examples{
prune.out = prune_archetypes(ACTION.out$C, ACTION.out$H)
G = build_ACTIONet(prune.out$H_stacked)
unification.out = unify_archetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
}
